requirements.txt
streamlit
langchain
python-dotenv
langchain_community
langchain-groq
validators
youtube-transcript-api==1.1.1
unstructured